Added abstract video codec
